refactor: useSolvedModal hooks > SolvedModal 폴더 이동
[#2]
