+++ packages/frontend/src/components/quiz/SolvedModal/index.ts
@@ -1,1 +1,2 @@
 export { SolvedModal } from "./SolvedModal";
+export { useSolvedModal } from "./useSolvedModal";
