+++ packages/frontend/src/components/quiz/index.ts
@@ -1,4 +1,4 @@
 export { default as QuizContent } from "./QuizContent";
 export { default as QuizLocation } from "./QuizLocation";
 export { default as CommandAccordion } from "./CommandAccordion";
-export { SolvedModal } from "./SolvedModal";
+export { SolvedModal, useSolvedModal } from "./SolvedModal";
